fixing Review model, service, dto

+++ Backend/DTO/Review/ReviewDTO.cs
@@ -2,4 +2,4 @@ using System;
 
 namespace Backend.DTO.Review;
 
-public record ReviewDTO(int Id, int UserId, int DockId, int Rating, string Content, DateTime CreatedAt, DateTime? UpdatedAt);
+public record ReviewDTO(int Id, int UserId, int DockId, double Rating, string Content, DateTime CreatedAt, DateTime? UpdatedAt);
